Fix bugs in query tagging for Snowflake.

+++ client/project/src/main/java/org/bsc/dcc/vcv/QueryRecord.java
@@ -61,7 +61,7 @@ public class QueryRecord {
 		this.tuples = tuples;
 	}
 	
-	public long getTuples() {
+	public int getTuples() {
 		return this.tuples;
 	}
